Create a file for the main Linter class
